Signals/pyCode/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.25.0
openpyxl>=3.0.0
pyarrow>=10.0.0
polars>=0.20.0
pyreadstat>=1.2.0